Extract Prompt Studio resizable input helper
